add product api
