+++ src/main/java/com/example/ecommerce/service/customer/CustomerResource.java
@@ -1,7 +1,7 @@
 package com.example.ecommerce.service.customer;
 
-import com.example.ecommerce.representation.request.customer.UserDetails;
-import com.example.ecommerce.representation.response.UserResponse;
+import com.example.ecommerce.service.representation.request.customer.UserDetails;
+import com.example.ecommerce.service.representation.response.UserResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
