+++ src/main/java/com/example/ecommerce/service/order/OrderResource.java
@@ -1,8 +1,8 @@
 package com.example.ecommerce.service.order;
 
 import com.example.ecommerce.exceptions.UserNotFoundException;
-import com.example.ecommerce.representation.request.order.OrderDetails;
-import com.example.ecommerce.representation.response.ItemsDetail;
+import com.example.ecommerce.service.representation.request.order.OrderDetails;
+import com.example.ecommerce.service.representation.response.ItemsDetail;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
