+++ src/main/java/com/example/ecommerce/service/products/ProductResource.java
@@ -1,7 +1,7 @@
 package com.example.ecommerce.service.products;
 
 import com.example.ecommerce.model.product.Inventory;
-import com.example.ecommerce.representation.request.products.ProductDetails;
+import com.example.ecommerce.service.representation.request.products.ProductDetails;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
