+++ src/main/java/com/example/ecommerce/service/representation/request/customer/UserDetails.java
@@ -1,5 +1,6 @@
-package com.example.ecommerce.representation.response;
+package com.example.ecommerce.service.representation.request.customer;
 
+import com.example.ecommerce.model.customer.User;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -10,6 +11,6 @@ import javax.xml.bind.annotation.XmlRootElement;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class UserResponse {
-    private String userId;
+public class UserDetails {
+    private User userData;
 }
